maybe ok

+++ eratosthenes.c
@@ -13,7 +13,7 @@ void Eratosthenes(BitArray_t array)
     unsigned long lenght = BA_size(array);
     BA_set_bit(array, 0, 1);
     BA_set_bit(array, 1, 1);
-    for (unsigned long i = 2; i < sqrt(lenght); i++)
+    for (unsigned long i = 2; i < (unsigned long) sqrt((unsigned long) lenght); i++)
     {
         for (unsigned long j = 2*i; j < lenght; j += i)
             BA_set_bit(array, j, 1);
